+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 - `sandstone/Heading` `font-style` to use oblique font instead of fake `italic`
 - `sandstone/Input` to not have initial focus with pointer when `type` prop is `'number'` or `'passwordnumber'`
 - `sandstone/Panel` to not reset scroll position by events from others
+- `sandstone/Panels.Header` to not show back button in the first panel
 - `sandstone/VideoPlayer.Video` to not start a new play before another one completes
 
 ## [1.4.4] - 2020-11-06